test: split instrumented tests from the e2e suite
Move the tests that need external infrastructure or secrets (the E2eSuite
with its helpers, plus NetworkConnectivityStressTest) into the
io.netbird.client.e2e package. The classic instrumented tests stay in the
root package and run again in CI via a secret-free instrumented-tests job
filtered with notPackage=io.netbird.client.e2e — safe for fork PRs.

The e2e package is executed from the private mobile-e2e repo, which selects
it with class=io.netbird.client.e2e.E2eSuite. New tests sort themselves:
anything placed in the e2e package is automatically excluded from the
public CI run, no exclude list to maintain.
