+++ app/src/androidTest/README.md
@@ -22,7 +22,7 @@ adb install -r -t app/build/outputs/apk/debug/app-debug.apk
 adb install -r -t app/build/outputs/apk/androidTest/debug/app-debug-androidTest.apk
 
 # 3. Run the test directly via adb
-adb shell am instrument -w -e class io.netbird.client.NetworkConnectivityStressTest \
+adb shell am instrument -w -e class io.netbird.client.e2e.NetworkConnectivityStressTest \
   io.netbird.client.test/androidx.test.runner.AndroidJUnitRunner
 ```
