+++ app/src/androidTest/java/io/netbird/client/e2e/ExitNodeRouteTest.java
@@ -1,4 +1,6 @@
-package io.netbird.client;
+package io.netbird.client.e2e;
+
+import io.netbird.client.MainActivity;
 
 import android.os.Bundle;
 import android.util.Log;
